chore: migrate from Babel to Rollup and refactor tests to use shared suite
