--- .babelrc
@@ -1,4 +1,0 @@
-{
-    "presets": ["@babel/preset-env"],
-    "plugins": ["@babel/plugin-transform-class-properties"]
-}
